feat!: remove legacy richtext (#705)
* feat!: remove legacy richtext

BREAKING CHANGE: `renderRichText` now uses `@storyblok/richtext` DX instead of the legacy schema resolve approach.

- [Release notes](https://github.com/storyblok/storyblok-js-client/releases)
- [Changelog](https://github.com/storyblok/storyblok-js-client/blob/main/changelog.md)
- [Commits](https://github.com/storyblok/storyblok-js-client/compare/v6.11.0...v7.0.0)

* refactor(tests): remove isRichTextEmpty tests following legacy richtext removal

This commit removes the tests for the `isRichTextEmpty` function, which is no longer relevant after the transition to the new `renderRichText` implementation using `@storyblok/richtext`. This aligns the test suite with the latest changes in the codebase.

* refactor(test): test only if `renderRichText` is wrapping correctly `richTextResolver`

* refactor(tests): simplify RichText e2e tests by removing unnecessary cases

* chore(deps): update various dependencies in pnpm-lock.yaml